Add in a bunch of #defines for packet types.
Show the packet type as a string, as well as a hex number, in the detail
display.  Show the packet type as a hex number in the summary display if
it's not a packet type we know about.

Don't put the "N on link, M captured" stuff in the PPP detail display,
as that's now in the frame detail display.

Handle IPv6, Appletalk, IPX, and Vines.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@79 f5534014-38df-0310-8fa8-9805f1628bb7