Don't preserve the read filter from file to file - you won't necessarily want to read the next file with the same filter that you used on the last file.
In the "File/Open" dialog box, parse the read filter before trying to
open the file, and if the parse fails, leave the dialog box up so the
user still has the filter and file name around and can try to fix the
problem.

Keep the compiled read filter attached to the "capture_file" structure,
so you don't have to reparse it on a "File/Reload".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@497 f5534014-38df-0310-8fa8-9805f1628bb7