"open_cap_file()" calls "close_cap_file()" if its attempt to open the capture file succeeds, so that it cleans out any state in the "capture_file" structure it was handed before filling it in with new state for the new capture file.
This means it destroys any read filter associated with the
"capture_file" structure it was handed, so the "rfcode" field must be
set *after* calling "open_cap_file()" but *before* calling
"read_cap_file()".  Do so for the "capture file (and possibly read
filter) supplied on the command line" case.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@520 f5534014-38df-0310-8fa8-9805f1628bb7