Dump the values of the "flags" and "ttl" fields of the CDP header.
Don't treat the CDP header as an in-memory data structure; that might
cause problems if it's not aligned on a 2-byte boundary.

Make the type and length fields of a TLV unsigned.

Correctly check for the end of the (captured part of the) frame.

Show most TLVs as "expandable" entries, where they expand into type,
length, and data entries.

Dissect "unknown" TLVs.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@573 f5534014-38df-0310-8fa8-9805f1628bb7