NBIPX packet type 3 appears to be the equivalent, in NBIPXland, of the NetBIOS Datagram Service in NBTland; a capture Gilbert sent had a pile of those packets containing what looked like SMB browser announcements, which are sent out as broadcast datagrams.  Label them as such, and treat them as such.
Might packet type 2 be the equivalent of the NetBIOS Session Service -
both of them contain SMBs, but the former is a connection-oriented
service (LLC I frames, presumably, in NBF, and TCP in NBT), and the
latter is a datagram-oriented service (LLC UI frames, presumably, in
NBF, and UDP in NBT)?  For now, we leave type 2 as "SMB (over NBIPX)",
but we might want to label it as "NetBIOS session" or whatever the
appropriate term is.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@574 f5534014-38df-0310-8fa8-9805f1628bb7