When reading a capture file, we can detect whether it's compressed or not, so it's OK to use "zlib" to read capture files, as it handles uncompressed files correctly.
When *writing* capture files, however, we can't detect automatically
whether the user wanted to write the file out as a compressed file or
not, so we should *NOT* use "zlib" until we add a flag to the API
specifying whether to write the file out as a compressed file or not.

Furthermore, the code in Ethereal that implements the "-S" flag depends
on being able to get the "FILE *" for a capture file being written, so
that it can "fflush()" it.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@703 f5534014-38df-0310-8fa8-9805f1628bb7