Have "do_capture()" take, as an argument, a pointer to the name of the file to which to write the capture; if it's NULL, create a temporary file and use that.
Have "-w" set a local variable, which starts out null, and, for "-k"
captures, call "do_capture()" and pass it that local variable as an
argument; this lets you do "-k" without "-w", which makes it use a
temporary file for the capture.

This means "run_capture()" no longer serves a useful purpose, as its
only caller is "do_capture()"; swallow it into "do_capture()".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@748 f5534014-38df-0310-8fa8-9805f1628bb7