The answer to the question
	(Is there a better way to force a parse to fail from arbitrary
	places in routines called by the parser?)

asked in an earlier checkin is "yes", which would've been obvious had I
seen the code that handles MAC addresses, as it returns NULL on an
error, and the YACC clause checks for a null return value and, if the
return value is null, uses YYERROR to make the parse fail.

Use that for IPv4 and IPv6 errors.

Also, use "dfilter_fail()" for the MAC address code.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@810 f5534014-38df-0310-8fa8-9805f1628bb7