For ONC RPC, when constructing conversations, use a null address as the destination address for calls and the source address of the reply - we should't require the server address to be the same for a call and reply, as they may not be on a multi-homed server (clients presumably check the XID only, or perhaps the XID and the port whence the reply came, although with TI-RPC I don't think they can check the port without checking the address as well).
This requires that the conversation code not assume that the source and
destination addresses for a given packet in a conversation have the same
type, so, when comparing addresses for equality, it must explicitly
check the address types.

In said code, also check the port numbers before we check the addresses
- testing ports is cheaper, as they're just integers, and there's
probably a decent chance that you won't see two conversations between
different pairs of hosts and the *same* pair of ports in a capture file,
so the cheaper port tests are probably decently likely to fail first.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1031 f5534014-38df-0310-8fa8-9805f1628bb7