Of the "zlib" functions we use that aren't in all versions of "zlib", "gzgets()" is the one most recently added; it was added in 1.0.9.
Check for it, rather than for a list of functions, when checking for
"zlib" support - if you check for N functions, and they're all there,
you get N "-lz"s added to the list of libraries with which to link.

Indicate in the README that "zlib" versions prior to 1.0.9 definitely
won't work.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1144 f5534014-38df-0310-8fa8-9805f1628bb7