Add some infrastructure for the use of Ethereal, so it can make a list of all the file types in which a file can be saved.
Giving each dumpable file type a routine that checks whether a file of a
given file type and encapsulation can be written lets us hoist some
checks into common code from out of the open routines.

If the "dump close" routine for a dump stream is NULL, have that mean
that there's no action that needs to be taken on a close by the code to
handle that file type; some file types don't need that, as they can be
written purely sequentially.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1200 f5534014-38df-0310-8fa8-9805f1628bb7