Add a check for "libtool".
Add the minimum version of {autoconf,automake,libtool} required to the
message printed if it's not found (having it check to make sure you have
at least that version is a bit more work).

Update the paths for "ftp.gnu.org" to reflect the fact that they now give
each package its own subdirectory.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1399 f5534014-38df-0310-8fa8-9805f1628bb7