If the attempt to execute Ethereal in a child process fails, have the child process send to the parent a message indicating why it failed, so that the parent can report that, and then exit.
If the attempt to create the child process to run Ethereal fails, pop up
a dialog box indicating that this happened.

Change the fork code a bit, to make it easier to substitute, on Win32
systems, code that does a "CreateProcess()" for the small chunk of code
that does the fork and exec.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1910 f5534014-38df-0310-8fa8-9805f1628bb7