Have tvbuff's keep track of cap_len and pkt_len ('length' and 'reported_length' in tvbuff terminology). This is implemented for TVBUFF_REAL and TVBUFF_SUBSET so far; support for TVBUFF_COMPOSITE is coming soon.
Throw either ReportedBoundsError or BoundsError.
A ReportedBoundsError is reported as "Malformed Frame" since the protocol
stated that a certain number of bytes should be available but they weren't.
A BoundsError is reported as a "Short Frame" since the snaplen was too short.

Register proto_short (BoundsError) and proto_malformed (ReportedBounds)
so searches can be made on "short" and "malformed".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@1965 f5534014-38df-0310-8fa8-9805f1628bb7