Use "progdlg_t *" rather than "void *" as the handle for a progress dialog box; that lets us do some type-checking, but we can still typedef it to an incompletely-defined structure to hide the implementation details from the caller.
Make "create_progress_dlg()" take, as an argument, the title to put in
the "stop the operation" button, and use "Stop" rather than "Cancel" if
stopping the operation doesn't undo all the work it's done.

Thaw the clist if we break out of a "read the file" operation, as we
freeze it before the operation.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@2121 f5534014-38df-0310-8fa8-9805f1628bb7