In the final scene of the horror movie, just when you think the monster is finally dead, and you're walking away, it springs up again and attacks.
It appears that the ss990915 version of Alexey Kuznetzov's libpcap patch
has some extra stuff in the per-packet header for some sort of SMP
debugging, and that SuSE Linux 6.3 picked it up.

Thus, even if a libpcap file has the modified magic number, we *still*
have to go through the usual heuristic hell to figure out what type of
file it is.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@2164 f5534014-38df-0310-8fa8-9805f1628bb7