- make-reg-dot
Do not assume that all dissectors are in the source directory.
This is the case for instance for packet-ncp2222.c which is
generated in the production directory (current).
This is particularly important when the generation is not
made in the source directory.

- ncp2222.py

Remove tabs in generated source to make the generated code
nicer and to make work the make-reg-dotc. Without that fix,
all NCP packet dissections dump core ...


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@2241 f5534014-38df-0310-8fa8-9805f1628bb7