It was silly of me to require that "forget_scrolled_window()" be called explicitly; the right thing to do was to have "remember_scrolled_window()" catch the "destroy" signal on the scrolled window widget, and have the handler for that signal forget the scrolled window.
Doing that obviates the need to have creators of scrolled windows keep
track of the windows they have and forget them when they're destroyed.

The signal for a "the window manager has requested that this window go
away" event is "delete_event", not "delete-event"; fix the
"gtk_signal_connect()" calls that were using "delete-event".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@2284 f5534014-38df-0310-8fa8-9805f1628bb7